fixed api views in exp layer and added views in exp

+++ model/marketplace/views.py
@@ -14,20 +14,37 @@ from .models import User, Carpool, Authenticator
 
 
 def success_response(status_code, detail='', auth_token=None, data=None):
-    assert isinstance(detail, str) and (auth_token is None or len(auth_token) == 64)
+    if auth_token and len(auth_token) != 64:
+        return JsonResponse({'detail': 'This authenticator is not valid.'}, status=400)
     response = {'detail': str(detail), 'auth_token': auth_token}
     if data:
-        if isinstance(data, dict) or isinstance(data, QueryDict):
-            response['data'] = json.dumps(data)
-        else:
+        if isinstance(data, dict) or isinstance(data, QueryDict) or isinstance(data, list):
             response['data'] = data
+        else:
+            try:
+                response['data'] = json.loads(data)
+            except ValueError:
+                return JsonResponse({'detail': 'This data is not valid.'}, status=400)
     return JsonResponse(response, status=status_code)
 
 
 def failure_response(status_code, detail='', auth_token=None):
-    assert isinstance(detail, str) and (auth_token is None or len(auth_token) == 64)
+    if auth_token and len(auth_token) != 64:
+        return JsonResponse({'detail': 'This authenticator is not valid.'}, status=400)
     response = {'detail': str(detail), 'auth_token': auth_token}
     return JsonResponse(response, status=status_code)
+
+
+def www_urlencoded_to_dict(body):
+    """
+    Converts raw x-www-form-urlencoded request data to mutable dict
+    This QueryDict instance is immutable error, must QueryDict.copy()
+    :param body:
+    :return: QueryDict or None
+    """
+    body = body.decode('utf-8')
+    data = QueryDict(body).copy()
+    return data
 
 
 def get_authenticator(auth_token=None, username=None):
@@ -71,7 +88,7 @@ def update_user(body, instance=None):
     :return:
     """
     if not isinstance(body, QueryDict) or not isinstance(body, dict):
-        return failure_response(status_code=400)
+        return failure_response(status_code=400, detail='Form bodies must be of dict type.')
     form = UserForm(body, instance=instance)
     if form.is_valid():
         user = form.save()
@@ -88,8 +105,10 @@ def update_user(body, instance=None):
 
 
 def update_carpool(body, instance=None):
-    if not isinstance(body, QueryDict) or not isinstance(body, dict) or 'auth_token' not in body:
-        return failure_response(status_code=400)
+    if not isinstance(body, QueryDict) or not isinstance(body, dict):
+        return failure_response(status_code=400, detail='Form bodies must be of dict type.')
+    if 'auth_token' not in body:
+        return failure_response(status_code=401, detail='This user is not logged in to perform this request.')
     form = CarpoolForm(body, instance=instance)
     if form.is_valid():
         carpool = form.save()
@@ -128,15 +147,21 @@ class UserView(View):
             if pk or username:
                 user = get_user(pk=pk) if pk else get_user(username=username)
                 if user:
-                    data = json.loads(serializers.serialize('json', [user, ]))
+                    data = {'users': json.loads(serializers.serialize('json', [user, ])), 'count': 1}
                     return success_response(status_code=200, data=data)
                 else:
                     return failure_response(status_code=404, detail='This user does not exist.')
             else:
-                users = User.objects.all()
+                users = None
+                print(bool(not request.GET))
+                if request.GET.get('carpool_owned'):
+                    users = User.objects.filter(carpool_owned=request.GET['carpool_owned'])
+                elif request.GET.get('carpool_joined'):
+                    users = User.objects.filter(carpool_joined=request.GET['carpool_joined'])
+                else:
+                    users = User.objects.all()
                 if users:
-                    data = json.loads(serializers.serialize('json', users))
-                    data['count'] = users.count()
+                    data = {'users': json.loads(serializers.serialize('json', users)), 'count': users.count()}
                     return success_response(status_code=200, data=data)
                 else:
                     return failure_response(status_code=404, detail='These users do not exist.')
@@ -148,8 +173,10 @@ class UserView(View):
         :return:
         """
         if request.method == 'POST':
-            body = request.body.decode('utf-8')
-            body = QueryDict(body)
+            if 'Content-Disposition: form-data' in request.body.decode('utf-8'):
+                return failure_response(status_code=400,
+                                        detail='Requests only accept x-www-form-urlencoded or json data.')
+            body = www_urlencoded_to_dict(request.body)
             if 'username' in body and 'password' in body:
                 if User.objects.filter(username=body['username']).exists():
                     return failure_response(status_code=409, detail='This username already exists.')
@@ -168,8 +195,10 @@ class UserView(View):
         :return:
         """
         if request.method == 'PUT':
-            body = request.body.decode('utf-8')
-            body = QueryDict(body)
+            if 'Content-Disposition: form-data' in request.body.decode('utf-8'):
+                return failure_response(status_code=400,
+                                        detail='Requests only accept x-www-form-urlencoded or json data.')
+            body = www_urlencoded_to_dict(request.body)
             user = None
             if 'username' in body and 'password' in body:
                 if User.objects.filter(username=body['username']).exists():
@@ -199,7 +228,7 @@ class UserView(View):
                 user = get_user(pk=pk) if pk else get_user(username=username)
             if user:
                 user.delete()
-                return success_response(status_code=204, detail='This user has been deleted.')
+                return success_response(status_code=204)
             else:
                 return failure_response(status_code=404, detail='This user does not exist.')
 
@@ -221,7 +250,7 @@ class CarpoolView(View):
             if pk:
                 carpool = get_carpool(pk=pk)
                 if carpool:
-                    data = json.loads(serializers.serialize('json', [carpool, ]))
+                    data = {'carpools': json.loads(serializers.serialize('json', [carpool, ])), 'count': 1}
                     return success_response(status_code=200, auth_token=auth_token, data=data)
                 else:
                     return failure_response(status_code=404, auth_token=auth_token,
@@ -229,48 +258,33 @@ class CarpoolView(View):
             else:
                 carpools = None
                 if request.GET.get('driver'):  # get carpool by driver id
-                    if not isinstance(request.GET.get('driver'), int):
-                        return failure_response(status_code=400, detail='This driver does not exist.',
-                                                auth_token=auth_token)
-                    carpools = Carpool.objects.filter(driver=request.GET.get('driver'))
+                    carpools = Carpool.objects.filter(driver=request.GET['driver'])
                 elif request.GET.get('driver_username'):  # get carpool by driver username
-                    if not isinstance(request.GET.get('driver_username'), str):
-                        return failure_response(status_code=400, detail='This driver does not exist.',
-                                                auth_token=auth_token)
-                    driver = get_user(username=request.GET.get('driver_username'))
+                    driver = get_user(username=request.GET['driver_username'])
                     carpools = Carpool.objects.filter(driver=driver.pk)
                 elif request.GET.get('passenger'):
-                    if not isinstance(request.GET.get('passenger'), int):
-                        return failure_response(status_code=400, detail='This passenger does not exist.',
-                                                auth_token=auth_token)
-                    carpools = Carpool.objects.filter(passengers__in=[request.GET.get('passenger')])
+                    carpools = Carpool.objects.filter(passengers__in=[request.GET['passenger']])
                 elif request.GET.get('passenger_username'):
-                    if not isinstance(request.GET.get('passenger_username'), str):
-                        return failure_response(status_code=400, detail='This passenger does not exist.',
-                                                auth_token=auth_token)
-                    passenger = get_user(username=request.GET.get('passenger_username'))
+                    passenger = get_user(username=request.GET['passenger_username'])
                     carpools = Carpool.objects.filter(passengers__in=[passenger.pk])
                 elif (request.GET.get('location_start_lat') and request.GET.get('location_start_lon')) and (
                             request.GET.get('location_end_lat') and request.GET.get('location_end_lon')):
-                    location_start_lat = request.GET.get('location_start_lat')
-                    location_start_lon = request.GET.get('location_start_lon')
-                    location_end_lat = request.GET.get('location_end_lat')
-                    location_end_lon = request.GET.get('location_end_lon')
-                    if isinstance(location_start_lat, float) and isinstance(location_start_lon, float) and isinstance(
-                            location_end_lat, float) and isinstance(location_end_lon, float):
-                        carpools = Carpool.objects.filter(
-                            location_start_lat__range=(location_start_lat - 0.5, location_start_lat + 0.5),
-                            location_start_lon__range=(location_start_lon - 0.5, location_start_lon + 0.5),
-                            location_end_lat__range=(location_end_lat - 0.5, location_end_lat + 0.5),
-                            location_end_lon__range=(location_end_lon - 0.5, location_end_lon + 0.5))
-                    else:
-                        return failure_response(status_code=400, detail='These locations do not exist.',
-                                                auth_token=auth_token)
-                elif not request.GET.keys():
+                    try:
+                        location_start_lat = float(request.GET['location_start_lat'])
+                        location_start_lon = float(request.GET['location_start_lon'])
+                        location_end_lat = float(request.GET['location_end_lat'])
+                        location_end_lon = float(request.GET['location_end_lon'])
+                    except ValueError:
+                        return failure_response(status_code=400, detail='Location coordinates must be floating point.')
+                    carpools = Carpool.objects.filter(
+                        location_start_lat__range=(location_start_lat - 0.5, location_start_lat + 0.5),
+                        location_start_lon__range=(location_start_lon - 0.5, location_start_lon + 0.5),
+                        location_end_lat__range=(location_end_lat - 0.5, location_end_lat + 0.5),
+                        location_end_lon__range=(location_end_lon - 0.5, location_end_lon + 0.5))
+                else:
                     carpools = Carpool.objects.all()
                 if carpools:
-                    data = json.loads(serializers.serialize('json', carpools))
-                    data['count'] = carpools.count()
+                    data = {'carpools': json.loads(serializers.serialize('json', carpools)), 'count': carpools.count()}
                     return success_response(status_code=200, auth_token=auth_token, data=data)
                 else:
                     return failure_response(status_code=404, auth_token=auth_token,
@@ -283,8 +297,10 @@ class CarpoolView(View):
         :return:
         """
         if request.method == 'POST':
-            body = request.body.decode('utf-8')
-            body = QueryDict(body)
+            if 'Content-Disposition: form-data' in request.body.decode('utf-8'):
+                return failure_response(status_code=400,
+                                        detail='Requests only accept x-www-form-urlencoded or json data.')
+            body = www_urlencoded_to_dict(request.body)
             if 'auth_token' not in body:
                 return failure_response(status_code=401, detail='This user is not logged in to perform this request.')
             return update_carpool(body=body)
@@ -297,8 +313,10 @@ class CarpoolView(View):
         :return:
         """
         if request.method == 'PUT':
-            body = request.body.decode('utf-8')
-            body = QueryDict(body)
+            if 'Content-Disposition: form-data' in request.body.decode('utf-8'):
+                return failure_response(status_code=400,
+                                        detail='Requests only accept x-www-form-urlencoded or json data.')
+            body = www_urlencoded_to_dict(request.body)
             if 'auth_token' not in body:
                 return failure_response(status_code=401, detail='This user is not logged in to perform this request.')
             auth_token = body['auth_token']
@@ -316,15 +334,17 @@ class CarpoolView(View):
         :return:
         """
         if request.method == 'DELETE':
-            body = request.body.decode('utf-8')
-            body = QueryDict(body)
+            if 'Content-Disposition: form-data' in request.body.decode('utf-8'):
+                return failure_response(status_code=400,
+                                        detail='Requests only accept x-www-form-urlencoded or json data.')
+            body = www_urlencoded_to_dict(request.body)
             if 'auth_token' not in body:
                 return failure_response(status_code=401, detail='This user is not logged in to perform this request.')
             auth_token = body['auth_token']
             carpool = get_carpool(pk=pk)
             if carpool:
                 carpool.delete()
-                return success_response(status_code=204, detail='This carpool has been deleted.', auth_token=auth_token)
+                return success_response(status_code=204, auth_token=auth_token)
             else:
                 return failure_response(status_code=404, detail='This carpool does not exist.', auth_token=auth_token)
 
@@ -344,7 +364,7 @@ class AuthenticationView(View):
             authenticator = None
             if username:
                 authenticator = get_authenticator(username=username)
-            if auth_token:
+            elif auth_token:
                 authenticator = get_authenticator(auth_token=auth_token)
             if authenticator:
                 return success_response(status_code=200, detail='This authenticator is valid.',
@@ -361,19 +381,20 @@ class AuthenticationView(View):
         :return:
         """
         if request.method == 'POST':
-            try:
-                assert 'username' in request.POST and 'password' in request.POST
-            except AssertionError:
+            if 'Content-Disposition: form-data' in request.body.decode('utf-8'):
                 return failure_response(status_code=400,
-                                        detail='This user is missing username or password.')
-            user = get_user(username=request.POST['username'])
+                                        detail='Requests only accept x-www-form-urlencoded or json data.')
+            body = www_urlencoded_to_dict(request.body)
+            if 'username' not in body or 'password' not in body:
+                return failure_response(status_code=400, detail='This request is missing required fields.')
+            user = get_user(username=body['username'])
             if not user:
                 return failure_response(status_code=400, detail='This user does not exist.')
-            authenticator = get_authenticator(username=request.POST['username'])
+            authenticator = get_authenticator(username=body['username'])
             if authenticator:
-                return failure_response(status_code=409, detail='User {} is already authenticated.'.format(
-                    request.POST['username']), auth_token=authenticator.auth_token)
-            elif hasher.verify(request.POST['password'], user.password):
+                return failure_response(status_code=409, detail='This user is already logged in.',
+                                        auth_token=authenticator.auth_token)
+            elif hasher.verify(body['password'], user.password):
                 auth_token = hmac.new(
                     key=settings.SECRET_KEY.encode('utf-8'),
                     msg=os.urandom(32),
@@ -381,16 +402,15 @@ class AuthenticationView(View):
                 ).hexdigest()
                 if Authenticator.objects.filter(auth_token=auth_token).exists():
                     return failure_response(status_code=409, detail='This auth_token already exists for another user.')
-                auth = AuthenticatorForm({'username': request.POST['username'],
+                auth = AuthenticatorForm({'username': body['username'],
                                           'auth_token': auth_token,
                                           'date_created': datetime.now()})
                 auth.save()
-                return success_response(status_code=201,
-                                        detail='Authenticator is successfully created for user {}.'.format(
-                                            request.POST['username']), auth_token=auth_token)
+                return success_response(status_code=201, auth_token=auth_token,
+                                        detail='This authenticator has been successfully created.')
             else:
                 return failure_response(status_code=400,
-                                        detail='Passwords for user {} do not match.'.format(request.POST['username']))
+                                        detail='These passwords do not match.')
 
     def delete(self, request, username=None, auth_token=None):
         """
@@ -406,10 +426,10 @@ class AuthenticationView(View):
             authenticator = None
             if username:
                 authenticator = get_authenticator(username=username)
-            if authenticator:
+            elif authenticator:
                 authenticator = get_authenticator(auth_token=auth_token)
             if authenticator:
                 authenticator.delete()
-                return success_response(status_code=204, detail='This authenticator is successfully deleted.')
+                return success_response(status_code=204)
             else:
                 return failure_response(status_code=404, detail='This authenticator does not exist.')
